+++ docs/CHANGELOG.md
@@ -12,6 +12,9 @@ Unreleased
 - Chg: Profiles embeddable on third party sites: expand to include full text - https://helpdesk.transition-space.org/conversation/291?folder_id=23
 - Fix: Calendar event wall entry for 1.18 - https://helpdesk.transition-space.org/conversation/1689?folder_id=23
 - Enh: Add admin access to user DM's - https://helpdesk.transition-space.org/conversation/1660?folder_id=23 
+- Enh: add a new Event in the transition module:
+  Each time the `members_map` table is updated (add, delete, change), update the User Profile field with internal name `Module::$profileFieldLatLngInternalName` with the new Lat/Lng.
+  If not empty, use `lat_user` and `lng_user`. If empty, use `lat_zip_city` and `lng_zip_city`
 
 2.0 (July 12, 2024)
 --------------------